requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pymongo[srv]==4.7.3
motor==3.4.0
python-dotenv==1.0.1
anthropic==0.28.0
openai==1.35.3
numpy==1.26.4
pydantic==2.7.4
pydantic-settings==2.3.4
httpx==0.27.0
